layout/style/test/chrome/test_moz_document_serialization.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=
-->
<head>
  <title>Test for Bug </title>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css"/>
  <style type="text/css" id="style"></style>
</head>
<body>
<pre id="test">
<script type="application/javascript">

var rules = [
  { rule: "@-moz-document url(http://www.example.com/) {}" },
  { rule: "@-moz-document url('http://www.example.com/') {}" },
  { rule: '@-moz-document url("http://www.example.com/") {}' },
  { rule: "@-moz-document url-prefix('http://www.example.com/') {}" },
  { rule: '@-moz-document url-prefix("http://www.example.com/") {}' },
  { rule: "@-moz-document domain('example.com') {}" },
  { rule: '@-moz-document domain("example.com") {}' },
  { rule: "@-moz-document regexp('http://www.w3.org/TR/\\d{4}/[^/]*-CSS2-\\d{8}/') {}" },
  { rule: '@-moz-document regexp("http://www.w3.org/TR/\\d{4}/[^/]*-CSS2-\\d{8}/") {}' },
];

SimpleTest.waitForExplicitFinish();

  var style = document.getElementById("style");
  var style_text = document.createTextNode("");
  style.appendChild(style_text);

  for (var i in rules) {
    var obj = rules[i];
    var rule = obj.rule;

    style_text.data = rule;
    is(style.sheet.cssRules.length, 1, "should have one rule");
    var ser1 = style.sheet.cssRules[0].cssText;
    if ("is_canonical" in obj) {
      is(ser1, rule, "rule '" + rule + "' should serialize to itself");
    }

    style_text.data = ser1;
    is(style.sheet.cssRules.length, 1, "should have one rule");
    var ser2 = style.sheet.cssRules[0].cssText;
    is(ser2, ser1,
        "parse+serialize for rule '" + rule + "' should be idempotent");
  }

  SimpleTest.finish();


</script>
</pre>
</body>
</html>
